backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai>=1.6.1
weaviate-client>=4.5.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
pypdf2==3.0.1
python-magic==0.4.27
aiofiles==23.2.1
sse-starlette==1.6.5
websockets==11.0.3
requests==2.31.0
numpy==1.24.3
tiktoken==0.5.2 